Elimino obras del home y las paso a Pabellón
