+++ src/pages/index.jsx
@@ -6,23 +6,18 @@ import SEO from '../components/seo';
 import Layout from '../components/layouts/Layout';
 import SocialMedia from '../components/sections/social-media';
 import Modalidades from '../components/sections/modalidades';
-import ListaObras from '../components/sections/obras';
 import '../scss/styles.scss';
 
 const titlePage = 'Inicio';
 
-const IndexPage = ({ data }) => {
-  const listaObras =
-    (data && data.allDatoCmsObra && data.allDatoCmsObra.edges) || [];
-
+const IndexPage = () => {
   return (
     <>
       <SEO title={titlePage} />
       <Layout imgOpt="home-videos/Home.jpg">
         <SocialMedia />
         <Modalidades />
-        <ListaObras obras={listaObras} />
-        <Link to="/obras" className="btn-obras">
+        <Link to="/pabellon" className="btn-obras">
           <span className="btn-obras__text">Ver todas las Obras</span>
         </Link>
       </Layout>
